backend: architecture of the app restructured uttils folder added to app_routes
