--- backend/app_main/app_dependancies_helpers_global_vars/dependencies.py
@@ -1,16 +1,0 @@
-from app_main.app_imports import AsyncSession, create_async_engine, db_url, AsyncGenerator, Any
-from app_main.app_dependancies_helpers_global_vars.app_logging import logger
-
-# ! Create an asynchronous engine
-engine = create_async_engine(db_url, future=True)
-
-
-# ! Dependency to get the database session
-async def get_db() -> AsyncGenerator[AsyncSession, Any]:
-	async with AsyncSession(engine) as session:
-		try:
-			yield session
-		except Exception as e:
-			await session.rollback()
-			logger.error(e)
-			raise e
