+++ backend/app_main/app_routes_blueprints/app_ai.py
@@ -1,6 +1,6 @@
 from app_main.app_imports import StreamingResponse, BaseModel, APIRouter, Annotated, Optional, List, File, UploadFile, \
 	Form, io, Image, JSONResponse
-from app_main.app_dependancies_helpers_global_vars.helpers import stream_text_gemma, stream_text_gemmini
+from app_main.app_routes_blueprints.uttils.helpers_ai import stream_text_gemma, stream_text_gemmini
 
 router = APIRouter(prefix="/v1/AI", tags=["AI-Models"])
