+++ backend/app_main/app_routes_blueprints/app_user.py
@@ -1,6 +1,6 @@
 from app_main.app_imports import APIRouter, Depends, datetime, AsyncSession
 from app_main.app_models.models import Tasks, User
-from app_main.app_dependancies_helpers_global_vars.dependencies import get_db
+from app_main.app_routes_blueprints.uttils.dependancies import get_db
 
 router = APIRouter(prefix="/v1/user", tags=["user"])
